+++ front/client/api/Community.js
@@ -1,15 +1,16 @@
 import ApiRequest from './ApiRequest'
 import request from 'superagent'
+import RequestUrls from '../constants/RequestUrls'
 
 export const createCommunity = (name, description) => {
   const communityParams = {
     name: name,
     description: description
   }
-  return ApiRequest.post('/communities', communityParams)
+  return ApiRequest.post(RequestUrls.communities, communityParams)
 }
 
 export const getCommunities = () => {
-  return request.get(`${Config.ApiEndPoint}/communities.json`)
+  return request.get(`${Config.ApiEndPoint}${RequestUrls.communities}`)
     .then(res => res.body)
 }
