+++ front/client/constants/RequestUrls.js
@@ -1,0 +1,3 @@
+export default {
+  "communities": "/communities"
+}
